+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1 - 2026-05-17
+
+- Hide gutter markers for resolved comment threads
+- Keep editor gutter markers limited to open line comments
+
 ## 0.2.0 - 2026-05-10
 
 - First public PatchTalk release